checkpoint: fix parsing iso timestamp
The non ISO formatted timestamp has a space between date and time, but
the ISO timestamp does not. The csv-parser was designed to fill fields based on
space separation, but that cannot work due to the different number of space in
case of timestamp. This patch first parses the timestamp, and assigns based on
the timestamp format later the proper fields.

Signed-off-by: Kokan <kokaipeter@gmail.com>