pragma-grammar: rename rules to indicate their differences based on EOL handling
The LL_EOL token should be consumed by the pragma rule before manipulating
the include stack. Two of the rules do that (@include and @requires),
however currently only @include does that and @requires not doing it is a bug.

This is a preparatory patch that makes it easier to change how @requires
behaves and to make the grammar a bit easier to understand.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>