packaging/debian: add support for Java based plugins [upstreamonly]
This patch adds support for packaging Java based plugins.  Please note that
the actual packaging of the java stuff will only happen in the build
environment has the required build dependencies and that these dependencies
are not pulled in by our Build-Depends line in control.

The thinking behind this is that some Debian versions support the required
stuff, others don't and we can't explicitly specify this in our
Build-Depends line.

If you want Java support, please install the packages as outlined in
dbld/images/helpers/packages.manifest file.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>