[BUG] Fix kafka deadlock
rdkafka internally maintains a message buffer. syslog-ng fills the
rdkafka buffer optimistically. When the rdkafka buffer becomes full,
typically the result of a kafka failure, or there is a kafka failure the
kafka dest fifo becomes full. When the kafka dest fifo is full messages
are dropped.

Seems the worker thread is only scheduled when there is work to do, i.e.
messages are inserted into the fifo. When the fifo is full messages are
never inserted.

When the worker thread is not scheduled it seems the periodic,
`_drain_responses`, is also not scheduled so `rd_kafka_poll` is never
called and the internal rdkafka buffer is full of delivery reports that
need to be handled.

The simple fix here is to move the periodic out of the worker and onto
the driver. I don't think `rd_kafka_poll` is very expensive so this
approach is probably fine. A better improvement would be to introduce a
core module solely to run periodics within a target deadline.