diskq: add compaction(yes/no) parameter
This patch introduces a new parameter for disk based buffers, for example:

    tcp("127.0.0.1" port(2001) disk-buffer(compaction(yes) disk-buf-size(10M)));

Note the "compaction" argument, it can either be "yes" or "no". With
compaction(yes), syslog-ng will prune the unused space in the LogMessage
representation, making the representation of the LogMessage more
compact. This is useful when syslog-ng uses a lot of name-value pairs during
processing that become unused when delivering to a destination.

Simply unsetting these values using the unset() rewrite operation does not
help, as the internal representation of a LogMessage will not release
the memory associated with these name-value pairs due to performance
reasons that help when syslog-ng is CPU bound.

In some cases however, the size of this overhead becomes significant (I've
heard of 4x raw message size in some cases), which becomes unbearable.

For these cases, the compaction will drop "unset" values, making the Logessage
representation smaller at the cost of some CPU time that is needed to
perform compaction.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>