nvtable: avoid having to lookup the same handle twice in nv_table_add_value()
Previously we were performing index based lookup for the same handle twice,

  1) to check if the NVTable already has an entry with the same handle.
  2) then if there was none, we were doing a lookup to come up with the
     index table location we needed to insert the new element

The 2nd is redundant and this patch gets rid of that.  It should improve
performance in cases where we are inserting new key-value pairs into a
LogMessage.  Of course it only matters more if the number of key-value pairs
is higher, otherwise the cost of the lookup is negligible.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>