nvtable: improve __find_index_entry()
__find_index_entry() contained a conditional to check if the handle to find
is _before_ all elements or _after_ all elements.  The first one proved not
to help from a performance standpoint, so dropping that as the loop after
the short-cut comes to the same conclusion without any iterations, as the
condition for the while loop will be false right from the start.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>