affile: move default PollFileChanges actions into event handlers
It looks better if all actions are placed into the event handler
(both for base and subclass).
I did not make it into virtual methods, because there is no need
to override default behaviour, but to extend a default implementation.

Order of execution is important, as a call to log_pipe_notify() can end
the lifetime of a PollFileChanges object.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>