affile: fix race condition of worker thread and file polling in main thread
PollMultilineFileChanges class's on_eof() handler can initiate a flush,
which must be followed by disabling further file polling
(as it is done in case of on_read() and on_file_move() by simply returning from check_file()).

But when a worker thread started to flush out the last multiline message in buffer,
after the on_eof() handler returned the PollFileChanges timer rearmed,
which will continue to poll file events.
In case a new message arrives into the file while the worker thread is running,
PollFileChanges schedules a worker job, which leads to an assertion failure.

However PollMultilineFileChanges::on_eof() handler is more complex than the other event handlers,
we cannot simply prvent rearming the timer.
It's internal logic decides if we should continue polling the file or not,
which should be seen on the base class level.

Signed-off-by: Gabor Nagy <gabor.nagy@balabit.com>