format-gelf: add --omit-empty-values
There are sources, which do not set the ${PID}, so it stays empty.
We use `$(format-json)` in `$(format-gelf)` with `int` type hinted
`${PID}`. If `${PID}` is empty, the whole `$(format-json)` call will
fail, and it will produce an empty message:

[2020-02-03T10:27:08.261055] Casting error; value='', type-hint='int32'
[2020-02-03T10:27:08.261116] Outgoing message; message=''

`_pid` is just and additional field in GELF, so it can be skipped:
https://docs.graylog.org/en/3.1/pages/gelf.html

I also had to add a default value to `short_message` and `host`,
as they are mandatory options, but it is a valid case, that `syslog-ng`
has `${HOST}` and `${MSG}` empty (although only in rare cases).

`-` was choosen, for simplicity.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>