build: remove -fcommon to reveal multiple definition issues
The -fcommon option hides (merges) multiple definition issues such as the
persist-tool global variable definition in the previous commit.
Such multi-definitions are not intentional in the syslog-ng codebase.

"The -fcommon places uninitialized global variables in a common block.
This allows the linker to resolve all tentative definitions of the same
variable in different compilation units to the same object."

In new GCC versions, -fno-common is the default.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>