python: option for building both Python 2 and Python 3 modules
This patch adds a configure option `--enable-python2-and-3`, which
builds syslog-ng with both Python 2 and Python 3 support enabled.

syslog-ng compiled with this option will start the Python 2 driver if
python() or python2() is specified in the configuration.
python3() is for Python 3 plugins.

python() currently means Python 2, but it can be changed or removed in the
future.

In case `--enable-python2-and-3` is _not_ used, the old behavior remains:
python() launches the driver syslog-ng was configured with.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>