python: detect conflicting Python modules
Python 2 and Python 3 can not be used in the same syslog-ng instance,
because plugins are currently loaded with RTLD_GLOBAL (g_module_open()),
and libpython 2 and 3 have conflicting symbols.

Loading libpython with RTLD_LOCAL is not supported:
https://bugs.python.org/issue36753
https://docs.python.org/3/whatsnew/3.8.html (Changes in the C API)

An alternative would be using dlmopen(), but currently it is not
really supported by gdb.

Note: Installing both v2 and v3 modules is possible because plugin
discovery is done with the G_MODULE_BIND_LAZY (RTLD_LAZY) flag.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>