template: rename LEVEL to SEVERITY while keeping compatible naming
Unfortunately, 20 years ago, syslog-ng choose the term "LEVEL" to refer to
the severity of a message, as that was the language used by syslog(3) manual
page.

Quoting the manpage:

       The priority argument is formed by ORing together a facility
       value and a level value (described below).  If no facility
       value is ORed into priority, then the default value set by
       openlog() is used, or, if there was no preceding openlog()
       call, a default of LOG_USER is employed.

However RFC3164 adopted the term "severity", quoting RFC3164:

       The number contained within these angle brackets is known as
       the Priority value and represents both the Facility and
       Severity as described below.

syslog-ng predates the choice in RFC3164, however I think it is highly time
to adopt the language of the RFC, especially as it is better than the
original.

See discussions in github PR #3115 for further details.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>