http: handle 'exceptions' during formatting HTTP headers
When formatting HTTP headers, there is a step when a signal is emitted
to collect the 'rest of the headers'. These headers are generated by
a slot (plugin). As these plugins can implemented also in Python, we
need to support error propagation.

Each error reported back from a slot execution are transformed to
critical or non-critical errors.

Non-critical errors are catched by caller (flush) and trying to send the
request even when the plugin has failed to append the header items.

Critical errors are ends in a LTR_NOT_CONNECTED state: syslog-ng won't
try to send the HTTP request when the slot cannot append the header items.

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>