set-severity: Fix setting a wrong severity with empty template
When a template evaulates as an empty string, strtol() returns 0 and
endptr[0] is NULL.  This result in a misclassification of the message
with a severity of 0, aka "emergency".

Fix this by ensuring that there is nothing left to be read AND the end
position is different from the start position.

Signed-off-by: Romain Tartière <romain@blogreen.org>