tls: fix error logging when key-file or cert-file is not set
Before this change we only logged the following error, when both were
not set:

"You have a TLS enabled source without a X.509 keypair.
Make sure you have tls(key-file() and cert-file()) options,
TLS handshake to this source will fail"

If only one is not set, or none of them is set,
we still want to log the error, because we need both of them to work
properly.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>