loggen: fix crash with invalid parameterization
../bin/loggen -S --active-connections=10 -I 120 caused crash in
loggen, because mandatory parameters were not passed.

The reason is even if a module cannot start, stop was still called. In
our specific case, thread_array was accessed in stop, even though it
was not initialized in start (because start returned earlier due to error).

The fix is that start returns boolean now. Stop is only called a
plugin is started.

Signed-off-by: Antal Nemes <antal.nemes.hu@gmail.com>