dbld: export EXTRA_FILES_DIR to prepare scripts
This patch adds a simple framework for pushing extra files to docker based
image builds, by introducing an EXTRA_FILES_DIR variable that points to a
newly created directory within /dbld/images/helpers, which is then ADDed to
the image by the dockerfiles.

We are going to use this in a subsequent patch to copy the debian/control
file to the container.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>