exec-prop-check: fix execution in devshell
In the `syslog-ng-devshell` docker image `exec_prop_check` fails,
because it does not correctly determines the ivykis related compile
commands. For some reason in the docker image the ivykis source dir
has one more `../` in it, which does not get removed by the regex
given.

Although, I did not find the reason behind the dir level anomaly,
I believe this change will not make false positive regex matches,
the regex is strict enough with the change, too.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>