packaging/debian: drop librdkafka version dependency [upstreamonly]
The reason we are dropping the >= dependency in our Build-Depends is
that older Debian/Ubuntu versions that we still want to build for
may have an old version of librdkafka and we want the build-deps to
be satisfied even in this case.

When this happens the configure script will refuse to use the installed
librdkafka version and build syslog-ng _without_ rdkafka support,
which is exactly what we want.

If the librdkafka version is good enough, we would build librdkafka
as well (e.g. debian sid for instance).

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>