affile: add complex return value to file_opener_open_fd
Instead of a simple boolean result, file_opener_open_fd will return
that the error was temporary or permanent. (Currently spurious paths
is the only permanent error.)
If the error is permanent we can do nothing about it, so we skip the
logwriter_reopen method too, thus there will be no further attempt to
open the file, and the log message will be skipped by the file
destination.

Signed-off-by: Laszlo Szemere <laszlo.szemere@oneidentity.com>