dbld: change docker build root to /dbld instead of /dbld/images
This patch relocates the root of the "docker build" operation, thereby
making it easier to share code between the building of the image
and the runtime execution of the container.

The follow directory schema are set up:
  - the entire dbld directory is copied into the image into /dbld (sans
    files excluded by .dockerignore)
  - the dbld directory is mounted over /dbld at runtime (e.g. any local
    changes make it into the runtime instance, no need to rebuild
    the image for this purpose)
  - dbld/helpers is removed and some stuff is moved into dbld/images
    (e.g. scripts used by the images) or right to the root of the
    dbld subdirectory (package lists for instance)
  - dbld/images/helpers/extra-files are moved to /dbld/extra-files

Paths to access the dbld files within the container (either build or
runtime) are changed to use /dbld.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>