dbld: add framework for controlling build options via featureflags
syslog-ng is build using different options on various platforms.  Some of
the platforms have dependencies for a specific feature whereas others don't.

Previously this was controlled using a combination of package lists
(packages.manifest that controlled what gets installed in a docker image),
configure options (in debian/rules and syslog-ng.spec) and conditionals
scattered in our rpm/deb packaging scripts.

This patch is intended to introduce a higher level concept "featureflags",
which is specified in a single location (dbld/build.manifest).  This then
controls:

  1) the installation of build dependencies (via Build-Depends and
     BuildRequires lines)
  2) the configure arguments
  3) the set of packages built

This patch changes the format of build.manifest, read the description of the
format at the beginning of that file as a comment block.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>