afinter: limit internal queue capacity
The internal source has a temporary source-side queue in addition to the
destination queues. This is because internal() messages can be created
anywhere, in any worker/dedicated threads, on any log path; and those
threads and paths should never be blocked/stopped just because something
is congested on the internal() log paths. The temporary queue is there to
overcome this problem, and for other performance reasons.

This commits sets a hard-coded limit for the internal queue.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>