rewrite: refactor rewrite expression grammar
The rewrite expr rules were not using the same conventions, this patch aims
to unify that _without_ changing behavior.

Some free() operations are pulled earlier and the return value is
always set through last_rewrite even where it was not strictly necessary.

This pattern however is a better pattern to be copy/pasted to future rewrite
expressions: makes it easier/trivial to add support for condition(),
which has been missed several times in the past (and is a subject to a subsequent patch).

This also addresses a couple of LogTemplate reference leaks.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>