logwriter: fix invalid memory access in proto clients
LogProtoClientOptions is just a reference in the LogProtoClient
class, owners are usually driver instances, and the ownership is not
transferred to the proto client.

This is true for all *Options in syslog-ng.

The lifetime of these *Options objects are clear:
The owner of the options object outlives the user.
For example, AFSocketDestDriver is freed only after its LogWriter/
LogProtoClient is deinitialized.

There are exceptions, of course. For example, the keep-alive()
functionality of network drivers saves a deinitialized but open LogWriter
instance in the "ReloadStore".
The reload mechanism creates a new driver instance (with new writer/proto
options), and restores the LogWriter instance from the reload store.

This involves a log_writer_set_options() call, that reinitializes the
LogWriter by setting all borrowed references (options, control pipe, etc.).

Unfortunately, the LogProtoClient object (which is owned by LogWriter) has
a borrowed proto_options field as well, so it has to be updated too,
not just the writer.

This is an old issue, but we've triggered a valgrind "invalid read" with
the following commit: 558ddd833.

Reproduction:

log {
  source { example-msg-generator(num(1)); };
  destination { syslog("localhost" transport("tcp") port(5555)); };
};

Reload syslog-ng with the above configuration.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>