slog: Remove 1500 limit and enhance the build system
The secure logging template can now process messages of arbitrary length.

The core functionality of the secure logging module is now build as a static library that is used by both the command line utilities and the dynamically loadable module. This change affects both the CMake and the autoconf builds.

Other small enhancements are also included in this changeset.

Signed-off-by: Airbus Commercial Aircraft <secure-logging@airbus.com>