dbld: fix dockerhub automated builds
I mistakenly moved the build hooks in a previous PR to "./dbld", whereas it
should have remained in "./dbld/images/hooks".

Also the cwd is set by dockerhub to dbld/images which is not what we want,
so change the script to the root of the source tree and invoke docker build
from there.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>