logmsg-deserialize: handle_change was effected only by last handle
The handle_changed should be true, if any handle required an update.

The state->handle_changed variable should track if the logmsg nvtable should be
updated due to any handle get assigned different value. The handles get their
value in the order they are requested from log_msg_registry. The log_msg_registry
is persistent while the syslog-ng is running (reload does not effect it), but on
restart the values are forgotten.

Signed-off-by: Kokan <kokaipeter@gmail.com>