slog: GLib for argument parsing for all command line utilities introduced
All command line utilities now use GLib argument parsing and provide useful help
messages if called with missing or incorrect arguments. The manual pages of
the utilities have been updated accordingly.

A new manual page for section 7 of the manual has been created. This manual page
will contain an introduction to secure logging and provide the user with
examples on how to use the secure logging module correctly, i.e. it will be a
small user guide. This manual page is still incomplete at the time of this commit
and will be updated in subsequent change sets.

Signed-off-by: Airbus Commercial Aircraft <secure-logging@airbus.com>