secret-storage: remove additional rlimit check
These asserts are not precise enough, being root is not enough to
determine if rlimit is restricted or not.

We need to check if we have `CAP_IPC_LOCK` capability. We had a run
with that, but we had to depend on `SYSLOG_NG_ENABLE_LINUX_CAPS`,
which is unfortunate in a test, because if we user does not have libcap
the tests are not run, also we will never notice, if the test stops
running for any reason.

In reality these asserts test a linux functionality, and to make it run
correctly it introduces too much complexity for no real usefulness.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>