scl/paloalto: drop unmatching messages in panos-parser()
Since it is pretty difficult to classify paloalto logs from content alone,
we need to perform most of the parsing to know if the log is in matching
format. This change restructures panos-parser() to always drop if the input
is invalid, by adding flags(drop-invalid) to the csv-parser() instances.

We also move the setting of $PROGRAM/$MESSAGE to the end, as if we are
changing them as the first step (as was originally in the patch), it
would also perform the change even if we are processing a non-paloalto
log that happens to be processed by the parser.

This patch also removes redundancies of parsing, e.g. previously all
header fields were parsed twice: first when grabbing the initial
fields and second when we process the individual log types.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>