stats: freq calculation overflow
The gint freq could hold the value 2^31-1, that later used in the code as:
```
timespec_add_msec(&timer->expires, freq * 1000);
```
While the timespec_add_msec last argument is a glong that could hold 1000*2^31-1,
the multiplication happens in the gint domain, which overflows yielding a
negative number. Causing the timer imidiate trigger.

Signed-off-by: Kokan <kokaipeter@gmail.com>