kafka: Evaluate the topic name and query/insert into hashmap
Here, we evaluate the template topic name as and when a message arrives at the
Kafka destination.

The process starts with quering the hash map to check if the topic is there.
If it is not present, we insert the topic. Otherwise, we use the topic from
the hash map. The process is carried out under the protection of a mutex lock.

Signed-off-by: Vivin Peris <vivinperis98@gmail.com>