scl/paloalto: sync with paloalto v9.1 formats
The original panos-parser() used a custom naming while PaloAlto publishes
names for the various fields it emits. This patch changes to the "official"
naming and also extends the format with the extensions added up to PanOS 9.1

NOTE: this is an incompatible change over the previous incarnation but
should be more maintainable going forward.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>