kafka: Working prototype which takes care of valid topic names
The code here consists of expansion of topic name when it is a template,
checking if it is in hashmap. Insert if it isn't. The evt_tag_str have
been update to account for the presence of template in topic name.
Also create hash table using g_hash_table_new_full to benefit from the
destruction functions provided.

Signed-off-by: Vivin Peris <vivinperis98@gmail.com>