afsocket-dest: secret-storage can call the async-init cb from another thread
After ControlServer is running in a separate thread, secret storage calls
the async init cb from a different thread.

For example when password protected SSL cert is used and the password is set
by the user via 'syslog-ng-ctl credentials', we have to ensure that the
`afsocket_dd_try_connect` is executed in the main thread (there is a
main-loop-assert and there are ivykis objects registered during
`dd_try_connect`).

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>