kafka: Modified grammar for fallback topic and wrote validation logic
Made allowance for fallback topic which would help to prevent loss of messages
in case topic name is invalid. The driver ensures that the topic and fallback
topic are both valid
The flow is as follows
Case 1 - Topic name isn't a template.
Here, we check if the topic name is valid with conditions as in the official
kafka logic code. If the topic name isn't valid, the driver isn't initialized.
Here, we don't need fallback topic even though the grammar allows for the
definition of the same since its not possible for a message publish failure
due to a VALID topic name to be fixed by another valid topic name.

Case 2 - Topic name is a template.
Here, we first check if the topic name is in hashmap.
In case it isn't, we check if it is a valid topic name and if it is a
valid topic name, then insert into hashmap. Else, we use the fallback topic
if defined. In case fallback topic isn't defined, then we return false as it
is not possible to deliver the message.

Signed-off-by: Vivin Peris <vivinperis98@gmail.com>