mainloop: fix dangling pointer issue
`current_configuration` could become a dangling pointer when it is assigned
to `old_config` and later `old_config` is deinited - for this reason
we need to set it NULL after it is assigned to `old_config`.

`current_configuration` can be accessed concurrently, so we need to protect
it with a lock.

Signed-off-by: Laszlo Budai <laszlo.budai@outlook.com>