lib/cfg-parser: fix error reporting to use the actual location of the token
When the error reporting format was improved to include the entire include
stack, we started using location information from the lexer's state, which
may not be entirely correct.

The lexer sometimes is advance to the token triggering the error (because
of LALR or simply because we are erroring on the first token of a rule), in
these cases using the location information in the lexer is incorrect,
and we should use the yylloc value as passed by the grammar.

This fixes cases where the caret showing the exact error location points
a bit past to the token that causes the error, for instance:

@version: 3.29

rewrite foo {
  subst ("", "", value("MESSAGE") type("pcre") flags("global store-matches"));
};

This was reported as:

3       rewrite foo {
4----->   subst ("", "", value("MESSAGE") type("pcre") flags("global store-matches"));
4----->                                                                            ^
5       };
6

Please note that the caret points to the closing parenthesis instead of
the incorrect flag. With the fix we get:

1       @version: 3.29
2
3       rewrite foo {
4----->   subst ("", "", value("MESSAGE") type("pcre") flags("global store-matches"));
4----->                                                      ^^^^^^^^^^^^^^^^^^^^^^
5       };
6

This would have saved us at least one bug-report where the user didn't understand
the error he was getting.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>