kafka: Added tests for the kafka topic parameter
Wrote unit tests for the _calculate_topic and _validate_topic_name methods to
 handle valid, and invalid topic names.
Created an internal header file for the functions to test and marked them
non static.

Refactored the kafka_dd_init method to take care of the case when topic name
is a literal. Also refactored the kafka_dd_validate_topic_name method to
remove dependency on the the driver object and modified the parametrized
test to take this into account.

Signed-off-by: Vivin Peris <vivinperis98@gmail.com>