filter/regex: fix not setting all of the named matches
The nametable which pcre provides can have gaps (begin_index=-1 || end_index=-1)
and in such cases those entries must be skipped and the next entry processed.
The issue was that the current implementation did only process the next name
entry if the current entry is not a gap. Witch works fine as long as there is no
named regex entry that is optional (see new UT cases for failing cases).

Signed-off-by: Kokan <kokaipeter@gmail.com>