python: use ack tracker factory in Python sources
The ack_tracker attribute can be used to request an ack tracker
instance in Python code. This attribute has to be set in the init()
method or earlier.

This sets an AckTrackerFactory, not a real ack tracker.

After the initialization of the source/fetcher, self.ack_tracker could
be overwritten with the real AckTracker instance in the future, if
needed.

For example:

from syslogng import LogSource
from syslogng import InstantAckTracker, ConsecutiveAckTracker, BatchedAckTracker

class TestSource(LogSource):
    def init(self, options):
        self.ack_tracker = InstantAckTracker(ack_callback=self.save_bookmark)
        self.ack_tracker = ConsecutiveAckTracker(ack_callback=self.save_bookmark)
        self.ack_tracker = BatchedAckTracker(timeout=5000, batch_size=100,
                                             batched_ack_callback=self.save_bookmarks)
        return True

    def save_bookmark(self, bookmark):
        pass

    def save_bookmarks(self, bookmarks):
        pass

    def run(self):
        pass

    def request_exit(self):
        pass

Signed-off-by: László Várady <laszlo.varady@protonmail.com>