python: deinit ack tracker before running Python deinit()
Python source users may receive acknowledgements after they deinitialize
their source. This is true in general, but in case of BatchedAckTracker,
an incomplete but large batch will be acked after the deinitialization.

This patch reduces the number of post-deinit acks in case of
BatchedAckTracker, so users have a chance to react to those acks while
their source is still running.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>